AVX-32997: Move SGs
