+++ aviatrix_ha.py
@@ -16,6 +16,7 @@ import boto3
 import botocore
 import version
 from api.login import login_to_controller
+from csp.sg import restore_security_group_access, temp_add_security_group_access
 from errors.exceptions import AvxError
 from csp.instance import get_controller_instance
 
@@ -633,45 +634,6 @@ def run_initial_setup(ip_addr, cid, ctrl_version):
         raise AvxError("Could not bring up the new controller to the "
                        "specific version")
     return False
-
-
-def temp_add_security_group_access(client, controller_instanceobj, api_private_access):
-    """ Temporarily add 0.0.0.0/0 rule in one security group"""
-    sgs = [sg_['GroupId'] for sg_ in controller_instanceobj['SecurityGroups']]
-    if api_private_access == "True":
-        return True, sgs[0]
-
-    if not sgs:
-        raise AvxError("No security groups were attached to controller")
-    try:
-        client.authorize_security_group_ingress(
-            GroupId=sgs[0],
-            IpPermissions=[{'IpProtocol': 'tcp',
-                            'FromPort': 443,
-                            'ToPort': 443,
-                            'IpRanges': [{'CidrIp': '0.0.0.0/0'}]}
-                           ])
-    except botocore.exceptions.ClientError as err:
-        if "InvalidPermission.Duplicate" in str(err):
-            return True, sgs[0]
-        print(str(err))
-        raise
-    return False, sgs[0]
-
-
-def restore_security_group_access(client, sg_id):
-    """ Remove 0.0.0.0/0 rule in previously added security group"""
-    try:
-        client.revoke_security_group_ingress(
-            GroupId=sg_id,
-            IpPermissions=[{'IpProtocol': 'tcp',
-                            'FromPort': 443,
-                            'ToPort': 443,
-                            'IpRanges': [{'CidrIp': '0.0.0.0/0'}]}
-                           ])
-    except botocore.exceptions.ClientError as err:
-        if "InvalidPermission.NotFound" not in str(err) and "InvalidGroup" not in str(err):
-            print(str(err))
 
 
 def handle_login_failure(priv_ip,
